+++ src/main/kotlin/me/nobaboy/nobaaddons/commands/debug/DebugCommands.kt
@@ -1,4 +1,4 @@
-package me.nobaboy.nobaaddons.commands
+package me.nobaboy.nobaaddons.commands.debug
 
 import com.mojang.brigadier.context.CommandContext
 import me.nobaboy.nobaaddons.api.DebugAPI
@@ -9,13 +9,9 @@ import me.nobaboy.nobaaddons.api.skyblock.SkyBlockAPI
 import me.nobaboy.nobaaddons.commands.internal.Command
 import me.nobaboy.nobaaddons.commands.internal.Group
 import me.nobaboy.nobaaddons.core.mayor.Mayor
-import me.nobaboy.nobaaddons.utils.MCUtils
 import me.nobaboy.nobaaddons.utils.TextUtils.buildText
 import me.nobaboy.nobaaddons.utils.TextUtils.toText
-import me.nobaboy.nobaaddons.utils.items.ItemUtils.isSkyBlockItem
-import me.nobaboy.nobaaddons.utils.items.ItemUtils.skyblockItem
 import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource
-import net.minecraft.component.DataComponentTypes
 import net.minecraft.text.MutableText
 import net.minecraft.text.Text
 import net.minecraft.util.Formatting
@@ -23,76 +19,29 @@ import kotlin.jvm.optionals.getOrNull
 
 @Suppress("unused")
 object DebugCommands : Group("debug") {
-	private fun MutableText.data(vararg items: Pair<String, Any?>) {
+	internal fun MutableText.data(vararg items: Pair<String, Any?>) {
 		items.forEach {
 			append(Text.literal("${it.first}: ").formatted(Formatting.BLUE))
-			append(Text.literal(it.second.toString()).formatted(Formatting.AQUA))
+			append(Text.literal(it.second.toString()).formatted(Formatting.GRAY))
 			append("\n")
 		}
 	}
 
-	private fun CommandContext<FabricClientCommandSource>.dumpInfo(vararg items: Pair<String, Any?>) {
+	internal fun CommandContext<FabricClientCommandSource>.dumpInfo(vararg items: Pair<String, Any?>) {
 		val text = buildText {
-			append("-".repeat(20).toText().formatted(Formatting.GRAY))
+			append("-".repeat(20).toText().formatted(Formatting.DARK_GRAY, Formatting.BOLD))
 			append("\n")
 			data(*items)
-			append("-".repeat(20).toText().formatted(Formatting.GRAY))
+			append("-".repeat(20).toText().formatted(Formatting.DARK_GRAY, Formatting.BOLD))
 		}
 		source.sendFeedback(text)
 	}
 
+	val item = ItemDebugCommands
+
 	val party = Command.command("party") {
 		executes {
 			PartyAPI.listMembers()
-		}
-	}
-
-	object Item : Group("item", executeRoot = true) {
-		override fun execute(ctx: CommandContext<FabricClientCommandSource>): Int {
-			val item = MCUtils.player!!.mainHandStack
-			if(item.isEmpty || !item.isSkyBlockItem) {
-				ctx.source.sendError(Text.literal("You aren't holding a valid SkyBlock item"))
-				return 0
-			}
-			val itemData = item.skyblockItem()
-			ctx.dumpInfo(
-				"Item ID" to itemData.id,
-				"UUID" to itemData.uuid,
-				"Created" to itemData.timestamp?.elapsedSince(),
-				"Reforge" to itemData.reforge,
-				"Rarity" to itemData.rarity,
-				"Recombobulated" to itemData.recombobulated,
-				"Stars" to itemData.stars,
-				"Enchants" to itemData.enchantments,
-				"Gemstones" to itemData.gemstones,
-				"Power scroll" to itemData.powerScroll,
-				"Donated to Museum" to itemData.donatedToMuseum,
-			)
-			return 0
-		}
-
-		val dumpNbt = Command.command("nbt") {
-			executes {
-				val item = MCUtils.player!!.mainHandStack
-				if(item.isEmpty) {
-					source.sendError(Text.literal("You aren't holding an item"))
-					return@executes
-				}
-				println(item.get(DataComponentTypes.CUSTOM_DATA))
-				source.sendFeedback(Text.literal("Dumped item NBT to game logs"))
-			}
-		}
-
-		val dumpLore = Command.command("lore") {
-			executes {
-				val item = MCUtils.player!!.mainHandStack
-				if(item.isEmpty || !item.contains(DataComponentTypes.LORE)) {
-					source.sendError(Text.literal("You aren't holding an item with lore"))
-					return@executes
-				}
-				println(item.get(DataComponentTypes.LORE)!!.lines)
-				source.sendFeedback(Text.literal("Dumped item lore to game logs"))
-			}
 		}
 	}