+++ src/main/kotlin/me/nobaboy/nobaaddons/config/NobaConfig.kt
@@ -30,4 +30,7 @@ class NobaConfig {
 
 	@SerialEntry
 	val qol: QOLConfig = QOLConfig()
+
+	@SerialEntry
+	val repo: RepoConfig = RepoConfig()
 }